pyproject.toml
[tool.poetry]
name = "home-inventory"
version = "0.1.0"
description = ""
authors = ["Dmitry Kostromin <kupec-k@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
qrcode = "^8.0"
fpdf2 = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
